+++ packages/credential-sdk/src/resolver/did/index.js
@@ -1,4 +1,5 @@
 export { default as DIDKeyResolver } from './did-key-resolver';
 export { default as DIDJWKResolver } from './did-jwk-resolver';
 export { default as DIDResolver } from './did-resolver';
+export { default as DIDResolverWithDIDReplacement } from './did-resolver-with-did-replacement';
 export { default as UniversalResolver } from './universal-resolver';