+++ packages/credential-sdk/src/types/blob/index.js
@@ -1,5 +1,11 @@
 import { TypedStruct, withProp } from '../generic';
-import { BlobId, CheqdBlobIdValue, DockBlobIdValue } from './id';
+import {
+  BlobId,
+  CheqdBlobIdValue,
+  CheqdMainnetBlobIdValue,
+  CheqdTestnetBlobIdValue,
+  DockBlobIdValue,
+} from './id';
 import Blob from './blob';
 
 export * from './id';
@@ -16,6 +22,16 @@ export class CheqdBlobWithId extends withProp(
   BlobWithId,
   'id',
   CheqdBlobIdValue,
+) {}
+export class CheqdTestnetBlobWithId extends withProp(
+  CheqdBlobWithId,
+  'id',
+  CheqdTestnetBlobIdValue,
+) {}
+export class CheqdMainnetBlobWithId extends withProp(
+  CheqdBlobWithId,
+  'id',
+  CheqdMainnetBlobIdValue,
 ) {}
 export class DockBlobWithId extends withProp(
   BlobWithId,