+++ packages/credential-sdk/src/types/did/document/ident-ref.js
@@ -4,10 +4,12 @@ import { NamespaceDid } from '../onchain/typed-did';
 export default class IdentRef extends withQualifier(TypedTuple) {
   static Qualifier = '';
 
+  static Did = NamespaceDid;
+
   static Ident = TypedString;
 
   static get Classes() {
-    return [NamespaceDid, this.Ident];
+    return [this.Did, this.Ident];
   }
 
   static random(did) {