+++ packages/credential-sdk/CHANGELOG.md
@@ -1,10 +1,22 @@
 # @docknetwork/credential-sdk
 
+## 0.42.0
+
+### Minor Changes
+
+- Bump up all packages
+
 ## 0.40.0
 
 ### Minor Changes
 
 - Export blob resolver
+
+## 0.39.0
+
+### Minor Changes
+
+- Add tests for all types of identifiers that can be received by the modules, use replacement resolver for DID documents
 
 ## 0.39.0